Wow. That took way too long to figure out.
Spent a copious amount of time figuring out that I did not have SQL
Express downloaded. Installed on the machine and changed the connection
string ing app.settings file to the proper server and added a clause
'TrustServerCertificate=True' that allowed me to connect to the DB.

Long term I probably need to install the certs, but this is just running
on my local machine.

I also found and replaced all the instances of the word 'Sweetener' for
the properly spelled 'Sweetner'.

Looking at why my order history is not populating next.