Situation: There was a time when I was updating my React app from using react-router-dom v5 to v6. useHistory() was removed and in turn useNavigate() was put in its place. The routing had changed from Switch to the Routes component and the individual Route component syntax had changed. After fixing it through out the app I realized that it was routed was working everywhere except my edit functinality.
Task:
 I was perplxed, because what was happening was that when I clicked on on the edit button the screen became blank, but the path in the URL was clearly visible. So because console warnings still bother me, I attempted to fix one involving passing depencies into the use effect. I felt distracted doing that so I told myseld to focus.

Action:
The routing syntax was changed so I did a little trial and error, I noticed that the edit path had a regex (regular expression) and I decided to remove it. That worked. And I think that was the real solution, but now when I attempted to make a form input change, state would immidiately clear my selection. It kept me really confused, and because I knew that state wasn't being updated, I knew it had to do something with managing state. Plug and play, after sometime I was still lost after copious research, I just started addressing the console warnings and miracoulsy it worked.

Result:
To be clear, I think the real reason was the regex in the path, but because I added depencies in one useEffect and not another, the state change couldn't populate correctly because the two components weren't able to communicate properly. The real solution was to focus on one issue at a time; however, thankfully my curiousity to fixing warning signs was the reason I found the fix. App was successfully updated to the newest version of react router and that is the big winner. The always evolving state of software..